add edge to edge, add interceptor for api key, clean architecture
