+++ app/src/main/java/com/krachkovsky/it_anews/data/models/NewsApiArticle.kt
@@ -1,0 +1,15 @@
+package com.krachkovsky.it_anews.data.models
+
+import com.google.gson.annotations.SerializedName
+
+data class NewsApiArticle(
+    val author: String?,
+    val content: String?,
+    val description: String?,
+    val publishedAt: String?,
+    @SerializedName("source")
+    val newsApiSource: NewsApiSource,
+    val title: String?,
+    val url: String,
+    val urlToImage: String?
+)
