+++ app/src/main/java/com/krachkovsky/it_anews/data/models/NewsApiSource.kt
@@ -1,0 +1,6 @@
+package com.krachkovsky.it_anews.data.models
+
+data class NewsApiSource(
+    val id: String?,
+    val name: String?
+)
