+++ app/src/main/java/com/krachkovsky/it_anews/domain/models/Article.kt
@@ -1,0 +1,12 @@
+package com.krachkovsky.it_anews.domain.models
+
+data class Article(
+    val author: String?,
+    val content: String?,
+    val description: String?,
+    val publishedAt: String?,
+    val source: Source,
+    val title: String?,
+    val url: String,
+    val urlToImage: String?
+)
