+++ app/src/main/java/com/krachkovsky/it_anews/domain/models/NewsEverything.kt
@@ -1,0 +1,7 @@
+package com.krachkovsky.it_anews.domain.models
+
+data class NewsEverything(
+    val articles: List<Article>,
+    val status: String,
+    val totalResults: Int
+)
