+++ app/src/main/java/com/krachkovsky/it_anews/domain/models/Source.kt
@@ -1,0 +1,6 @@
+package com.krachkovsky.it_anews.domain.models
+
+data class Source(
+    val id: String?,
+    val name: String?
+)
