--- app/src/main/java/com/krachkovsky/it_anews/extention/RecyclerViewExtensions.kt
@@ -1,23 +1,0 @@
-package com.krachkovsky.it_anews.extention
-
-import android.graphics.Rect
-import android.view.View
-import androidx.recyclerview.widget.RecyclerView
-
-fun RecyclerView.addHorizontalSpaceDecoration(space: Int) {
-    addItemDecoration(
-        object : RecyclerView.ItemDecoration() {
-            override fun getItemOffsets(
-                outRect: Rect,
-                view: View,
-                parent: RecyclerView,
-                state: RecyclerView.State
-            ) {
-                val position = parent.getChildAdapterPosition(view)
-                if (position != 0 && position != parent.adapter?.itemCount) {
-                    outRect.top = space
-                }
-            }
-        }
-    )
-}
