--- app/src/main/java/com/krachkovsky/it_anews/models/NewsEverything.kt
@@ -1,7 +1,0 @@
-package com.krachkovsky.it_anews.models
-
-data class NewsEverything(
-    val articles: List<Article>,
-    val status: String,
-    val totalResults: Int
-)
