--- app/src/main/java/com/krachkovsky/it_anews/models/Source.kt
@@ -1,6 +1,0 @@
-package com.krachkovsky.it_anews.models
-
-data class Source(
-    val id: String?,
-    val name: String?
-)
