+++ app/src/main/java/com/krachkovsky/it_anews/presentation/adapter/LoadingViewHolder.kt
@@ -1,0 +1,6 @@
+package com.krachkovsky.it_anews.presentation.adapter
+
+import androidx.recyclerview.widget.RecyclerView
+import com.krachkovsky.it_anews.databinding.NewsLoadingBinding
+
+class LoadingViewHolder(binding: NewsLoadingBinding) : RecyclerView.ViewHolder(binding.root)
