--- app/src/main/java/com/krachkovsky/it_anews/retrofit/NewsApi.kt
@@ -1,20 +1,0 @@
-package com.krachkovsky.it_anews.retrofit
-
-import com.krachkovsky.it_anews.models.NewsEverything
-import com.krachkovsky.it_anews.util.Constants.API_KEY
-import retrofit2.http.GET
-import retrofit2.http.Query
-
-interface NewsApi {
-    @GET("everything")
-    suspend fun getAllNews(
-        @Query("q")
-        q: String = "Belarus",
-        @Query("page")
-        page: Int = 1,
-        @Query("pageSize")
-        pageSize: Int,
-        @Query("apiKey")
-        apiKey: String = API_KEY
-    ): NewsEverything
-}
