--- app/src/main/java/com/krachkovsky/it_anews/retrofit/NewsApiFactory.kt
@@ -1,26 +1,0 @@
-package com.krachkovsky.it_anews.retrofit
-
-import com.krachkovsky.it_anews.util.Constants.BASE_URL
-import okhttp3.OkHttpClient
-import okhttp3.logging.HttpLoggingInterceptor
-import retrofit2.Retrofit
-import retrofit2.converter.gson.GsonConverterFactory
-import retrofit2.create
-
-class NewsApiFactory {
-
-    private val retrofit by lazy(LazyThreadSafetyMode.NONE) {
-        val logging = HttpLoggingInterceptor()
-        logging.setLevel(HttpLoggingInterceptor.Level.BODY)
-        val client = OkHttpClient.Builder()
-            .addInterceptor(logging)
-            .build()
-        Retrofit.Builder()
-            .baseUrl(BASE_URL)
-            .addConverterFactory(GsonConverterFactory.create())
-            .client(client)
-            .build()
-    }
-
-    fun provideApi() = retrofit.create<NewsApi>()
-}
