--- app/src/main/java/com/krachkovsky/it_anews/ui/NewsMainActivity.kt
@@ -1,18 +1,0 @@
-package com.krachkovsky.it_anews.ui
-
-import android.graphics.Color
-import android.os.Bundle
-import androidx.appcompat.app.AppCompatActivity
-import androidx.core.view.WindowCompat
-import com.krachkovsky.it_anews.R
-
-class NewsMainActivity : AppCompatActivity(R.layout.activity_main) {
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-
-        WindowCompat.setDecorFitsSystemWindows(window, false)
-        window.statusBarColor = Color.TRANSPARENT
-        window.navigationBarColor = Color.TRANSPARENT
-    }
-}
