--- app/src/main/java/com/krachkovsky/it_anews/ui/adapter/ErrorViewHolder.kt
@@ -1,6 +1,0 @@
-package com.krachkovsky.it_anews.ui.adapter
-
-import androidx.recyclerview.widget.RecyclerView
-import com.krachkovsky.it_anews.databinding.NewsErrorBinding
-
-class ErrorViewHolder(binding: NewsErrorBinding) : RecyclerView.ViewHolder(binding.root)
