--- app/src/main/java/com/krachkovsky/it_anews/ui/adapter/LoadingViewHolder.kt
@@ -1,6 +1,0 @@
-package com.krachkovsky.it_anews.ui.adapter
-
-import androidx.recyclerview.widget.RecyclerView
-import com.krachkovsky.it_anews.databinding.NewsLoadingBinding
-
-class LoadingViewHolder(binding: NewsLoadingBinding) : RecyclerView.ViewHolder(binding.root)
