--- app/src/main/java/com/krachkovsky/it_anews/util/Constants.kt
@@ -1,6 +1,0 @@
-package com.krachkovsky.it_anews.util
-
-object Constants {
-    const val API_KEY = "08e6c887e888450baeec6d57ba241fbe"
-    const val BASE_URL = "https://newsapi.org/v2/"
-}
